WIP: initial version of Backward MHA test using C++ API
